Moved API call to service
Moved models to be with API relevant API service
