+++ src/services/torchbearer/torchbearer.service.ts
@@ -1,0 +1,13 @@
+import axios from "axios";
+
+class TorchbearerService {
+  apiUrl: string = "https://www.unionlarp.co.uk/torchbearer";
+
+  constructor() {}
+
+  getSkillsData() {
+    return axios.get(`${this.apiUrl}/gettorchbearerskills`);
+  }
+}
+
+export const torchbearerService = new TorchbearerService();
